Updates lidar configuration and messages
Updates the lidar configuration process and improves
logging for better clarity.

Increases the angle offset range and modifies the success
message to accurately reflect the configuration status.

+++ src/lidar/AbstractLidarHelper.h
@@ -41,7 +41,7 @@ protected:
   string comFile;
 
   bool reversed;
-  int8_t angleOffset;
+  int16_t angleOffset;
   int16_t excludeLowerThanMm;
   int16_t excludeGreaterThanMm;
